services/django-auth/pyproject.toml
[project]
name = "django-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faker>=40.1.2",
    "hypothesis>=6.151.5",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
]
